refactor: delete root thread projection shell
