+++ backend/monitor/infrastructure/read_models/thread_workbench_read_service.py
@@ -8,8 +8,8 @@ from datetime import UTC, datetime
 from typing import Any
 
 from backend.avatar_urls import avatar_url
-from backend.thread_projection import canonical_owner_threads
 from backend.thread_runtime.convergence import converge_owner_thread_runtime, summarize_owner_thread_runtime
+from backend.thread_runtime.projection import canonical_owner_threads
 from core.runtime.middleware.monitor import AgentState
