--- backend/thread_projection.py
@@ -1,3 +1,0 @@
-"""Compatibility shell for thread runtime projection helpers."""
-
-from backend.thread_runtime.projection import *  # noqa: F403
