+++ backend/user_sandbox_reads.py
@@ -7,7 +7,7 @@ from collections import Counter
 from typing import Any
 
 from backend.avatar_urls import avatar_url
-from backend.thread_projection import canonical_owner_threads
+from backend.thread_runtime.projection import canonical_owner_threads
 from backend.virtual_threads import is_virtual_thread_id
 from sandbox.recipes import default_recipe_id, normalize_recipe_snapshot, provider_type_from_name
 from storage.models import map_sandbox_state_to_display_status
