+++ backend/web/services/sandbox_service.py
@@ -13,7 +13,7 @@ from backend import sandbox_runtime_mutations as _sandbox_runtime_mutations
 from backend import sandbox_runtime_reads as _sandbox_runtime_reads
 from backend import sandbox_thread_resources as _sandbox_thread_resources
 from backend.avatar_urls import avatar_url
-from backend.thread_projection import canonical_owner_threads
+from backend.thread_runtime.projection import canonical_owner_threads
 from backend.virtual_threads import is_virtual_thread_id
 from backend.web.core.config import LOCAL_WORKSPACE_ROOT, SANDBOXES_DIR
 from sandbox.config import SandboxConfig
